requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.5
django-cache-url==3.4.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
environs==9.5.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
marshmallow==3.20.1
multidict==6.0.4
openai==0.28.1
packaging==23.1
psycopg==3.1.11
psycopg-binary==3.1.11
psycopg2-binary==2.9.8
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
whitenoise==6.5.0
yarl==1.9.2